requirement.txt
Requirements:

Python 

Pytorch	1.8.1

PyQt5

tensorflow

matplotlib 3.4.1

numpy 1.20.2
